chore: rename to style/confirm.tsx
