+++ components/modal/style/confirm.tsx
